2.0.0
1. Cus dapat login menggunakan member umum (free pass)
2. Cus tidak bisa lagi ketik kode member manual untuk login Ikiosk
3. Cegatan member login di 2 station
4. Backup zip kalau gk ad trx gk keluar error
5. Fix double scan pembayaran

+++ src/pages/mainMenu/pembelianPembayaran/belanja/KasirPembayaran.jsx
@@ -21,6 +21,7 @@ import {
 } from "../../../../services/redux/dtAllInputtedItemReducer";
 import Modal from "../../../../components/Modal";
 import {
+  AESEncrypt,
   passToChar,
   PAYMENT_KEY,
   PIC_API_KEY,
@@ -35,6 +36,7 @@ import {
 import Lottie from "lottie-react";
 import { toggleMemberMerah } from "../../../../services/redux/memberReducer";
 import {
+  deleteTempMemberFromAPI,
   errorLog,
   generateTXTError,
   sendErrorLogWithAPI,
@@ -102,6 +104,7 @@ function KasirPembayaran() {
   const glUserModul = useSelector((state) => state.glUser.userModul);
   const glStationModul = useSelector((state) => state.glUser.stationModul);
   const glLougoutApp = useSelector((state) => state.glCounter.glLogOutLimitApp);
+  const glIpModul = useSelector((state) => state.glDtIp.dtIp);
   const appVersion = packageJson.version;
 
   const generateDocCode = (noTransaksi) => {
@@ -138,21 +141,43 @@ function KasirPembayaran() {
     }
   };
 
-  const handleNavigate = () => {
-    console.log("handlenavigate", memberMerah);
-    if (memberMerah) {
-      navigate("/");
-      dispatch(addDtTimeStart(""));
-      dispatch(removeAllItems());
-      setOpenModalAlert(false);
-      setOpenModalPayment(false);
+  const handleNavigate = async () => {
+    setLoading(true);
+    const doDeleteTempMemberFromAPI = await deleteTempMemberFromAPI(
+      URL_GATEWAY,
+      userDt["memberID"],
+      glIpModul,
+      glStationModul,
+      glRegistryDt
+    );
+
+    if (doDeleteTempMemberFromAPI.status === true) {
+      if (memberMerah) {
+        navigate("/");
+        dispatch(addDtTimeStart(""));
+        dispatch(removeAllItems());
+        setOpenModalAlert(false);
+        setOpenModalPayment(false);
+      } else {
+        navigate("/");
+        dispatch(addDtTimeStart(""));
+        dispatch(removeAllItems());
+        setOpenModalAlert(false);
+        setOpenModalPayment(false);
+        dispatch(toggleMemberMerah());
+      }
     } else {
-      navigate("/");
-      dispatch(addDtTimeStart(""));
-      dispatch(removeAllItems());
-      setOpenModalAlert(false);
-      setOpenModalPayment(false);
-      dispatch(toggleMemberMerah());
+      if (
+        doDeleteTempMemberFromAPI.message ===
+        "Network doDeleteTempMemberFromAPI"
+      ) {
+        setMsg("Gagal Terhubung Dengan Gateway");
+      } else {
+        setMsg(doDeleteTempMemberFromAPI.message);
+      }
+
+      setLoading(false);
+      setOpenModalAlert(true);
     }
   };
 
@@ -168,7 +193,7 @@ function KasirPembayaran() {
 
       const setNewTimeout = () => {
         newTimeoutId = setTimeout(async () => {
-          handleNavigate();
+          await handleNavigate();
         }, 120 * 1000);
       };
 
@@ -271,9 +296,13 @@ function KasirPembayaran() {
     if (e.key === "Enter") {
       setLoading(true);
       if (openModalPayment === false) {
-        setInputValue("");
-        setSelectedPayment("");
-        setLoading(false);
+        if (selectedPayment !== "") {
+          setInputValue("");
+        } else {
+          setInputValue("");
+          setSelectedPayment("");
+          setLoading(false);
+        }
       } else {
         setOpenModalPayment(false);
         let totalJualHDR =
@@ -578,13 +607,13 @@ function KasirPembayaran() {
         open={openModalAlert}
         successAlert={alertSucc}
         landscape={isLandscape}
-        onClose={() => {
+        onClose={async () => {
           setOpenModalAlert(false);
           setInputValue("");
           setAlertSucc(false);
           setGetPointGift(false);
           setAlertInfo(false);
-          alertSucc ? handleNavigate() : navigate("/kasirSelfService");
+          alertSucc ? await handleNavigate() : navigate("/kasirSelfService");
         }}
       >
         <div className="text-center">
